feat: criado TipoProdutoEnum e tratativas de erro para ele. Criado endpoint para cadastro de Produto

+++ comandaRestaurante/src/main/java/com/projetoweb4/comandaRestaurante/dto/domain/TipoProdutoDtoDetalhar.java
@@ -3,7 +3,7 @@ package com.projetoweb4.comandaRestaurante.dto.domain;
 import com.projetoweb4.comandaRestaurante.entity.domain.TipoProduto;
 
 public record TipoProdutoDtoDetalhar(
-		Long id,
+		Short id,
 		String nome
 		) {
